[MIG] mrp_bom_line_ref: Migration to 12.0
