+++ mrp_bom_line_ref/README.rst
@@ -1,0 +1,71 @@
+======================
+MRP BOM Line Reference
+======================
+
+.. !!!!!!!!!!!!!!!!!!!!!!!!!!!!!!!!!!!!!!!!!!!!!!!!!!!!
+   !! This file is generated by oca-gen-addon-readme !!
+   !! changes will be overwritten.                   !!
+   !!!!!!!!!!!!!!!!!!!!!!!!!!!!!!!!!!!!!!!!!!!!!!!!!!!!
+
+.. |badge1| image:: https://img.shields.io/badge/maturity-Beta-yellow.png
+    :target: https://odoo-community.org/page/development-status
+    :alt: Beta
+.. |badge2| image:: https://img.shields.io/badge/licence-AGPL--3-blue.png
+    :target: http://www.gnu.org/licenses/agpl-3.0-standalone.html
+    :alt: License: AGPL-3
+.. |badge3| image:: https://img.shields.io/badge/github-qrtl%2Fqrtl--addons-lightgray.png?logo=github
+    :target: https://github.com/qrtl/qrtl-addons/tree/12.0/mrp_bom_line_ref
+    :alt: qrtl/qrtl-addons
+
+|badge1| |badge2| |badge3| 
+
+This module provides the following functions:
+
+* Adds `Reference` field in BOM line (expected to be used to keep mapping reference for parts in circuit diagram).
+
+This module depends on ``mrp`` module.
+
+**Table of contents**
+
+.. contents::
+   :local:
+
+Installation
+============
+
+No special instructions.
+
+Usage
+=====
+
+You can fill in Reference Number in a Bom Line record in the form view of "Bill of Material".
+
+Bug Tracker
+===========
+
+Bugs are tracked on `GitHub Issues <https://github.com/qrtl/qrtl-addons/issues>`_.
+In case of trouble, please check there if your issue has already been reported.
+If you spotted it first, help us smashing it by providing a detailed and welcomed
+`feedback <https://github.com/qrtl/qrtl-addons/issues/new?body=module:%20mrp_bom_line_ref%0Aversion:%2012.0%0A%0A**Steps%20to%20reproduce**%0A-%20...%0A%0A**Current%20behavior**%0A%0A**Expected%20behavior**>`_.
+
+Do not contact contributors directly about support or help with technical issues.
+
+Credits
+=======
+
+Authors
+~~~~~~~
+
+* Rooms For (Hong Kong) Limited T/A OSCG
+
+Contributors
+~~~~~~~~~~~~
+
+* Quartile Limited <https://www.quartile.co/>
+
+Maintainers
+~~~~~~~~~~~
+
+This module is part of the `qrtl/qrtl-addons <https://github.com/qrtl/qrtl-addons/tree/12.0/mrp_bom_line_ref>`_ project on GitHub.
+
+You are welcome to contribute.
