+++ mrp_bom_line_ref/i18n/ja_JP.po
@@ -4,10 +4,10 @@
 #
 msgid ""
 msgstr ""
-"Project-Id-Version: Odoo Server 9.0e\n"
+"Project-Id-Version: Odoo Server 12.0\n"
 "Report-Msgid-Bugs-To: \n"
-"POT-Creation-Date: 2016-06-10 05:46+0000\n"
-"PO-Revision-Date: 2016-06-10 05:46+0000\n"
+"POT-Creation-Date: 2021-12-09 06:31+0000\n"
+"PO-Revision-Date: 2021-12-09 06:31+0000\n"
 "Last-Translator: <>\n"
 "Language-Team: \n"
 "MIME-Version: 1.0\n"
@@ -16,12 +16,11 @@ msgstr ""
 "Plural-Forms: \n"
 
 #. module: mrp_bom_line_ref
-#: model:ir.model.fields,field_description:mrp_bom_line_ref.field_mrp_bom_line_ref
-msgid "Reference Number"
-msgstr "参照番号"
+#: model:ir.model,name:mrp_bom_line_ref.model_mrp_bom_line
+msgid "Bill of Material Line"
+msgstr "部品表明細"
 
 #. module: mrp_bom_line_ref
-#: model:ir.model,name:mrp_bom_line_ref.model_mrp_bom_line
-msgid "mrp.bom.line"
-msgstr "mrp.bom.line"
-
+#: model:ir.model.fields,field_description:mrp_bom_line_ref.field_mrp_bom_line__ref
+msgid "Reference Number"
+msgstr "参照番号"
