+++ mrp_bom_line_ref/models/__init__.py
@@ -16,10 +16,4 @@
 #    You should have received a copy of the GNU Affero General Public License
 #    along with this program.  If not, see <http://www.gnu.org/licenses/>.
 
-from openerp import models, fields, api, _
-
-
-class MrpBomLine(models.Model):
-    _inherit = 'mrp.bom.line'
-
-    ref = fields.Char('Reference Number')
+from . import mrp
