+++ mrp_bom_line_ref/readme/DESCRIPTION.rst
@@ -1,0 +1,5 @@
+This module provides the following functions:
+
+* Adds `Reference` field in BOM line (expected to be used to keep mapping reference for parts in circuit diagram).
+
+This module depends on ``mrp`` module.
