+++ mrp_bom_line_ref/readme/INSTALL.rst
@@ -1,0 +1,1 @@
+No special instructions.
