+++ mrp_bom_line_ref/readme/USAGE.rst
@@ -1,0 +1,1 @@
+You can fill in Reference Number in a Bom Line record in the form view of "Bill of Material".
